changes based on hanzei's code review
